Rename getComponentsToUse to getComponentsFromServerlessFile
