+++ src/utils/components/index.js
@@ -1,9 +1,9 @@
 const generateInstanceId = require('./generateInstanceId')
-const getComponentsToUse = require('./getComponentsToUse')
 const getInstanceId = require('./getInstanceId')
 const getExecutedComponents = require('./getExecutedComponents')
 const getComponent = require('./getComponent')
 const getComponentFunctions = require('./getComponentFunctions')
+const getComponentsFromServerlessFile = require('./getComponentsFromServerlessFile')
 const getComponentsFromStateFile = require('./getComponentsFromStateFile')
 const getComponentsToRemove = require('./getComponentsToRemove')
 const executeComponent = require('./executeComponent')
@@ -11,9 +11,9 @@ const generateContext = require('./generateContext')
 
 module.exports = {
   generateInstanceId,
-  getComponentsToUse,
   getInstanceId,
   getComponentFunctions,
+  getComponentsFromServerlessFile,
   getComponentsFromStateFile,
   getComponentsToRemove,
   getExecutedComponents,
